azurecloud/backend/azure-functions/orchestrator/function_app.py
"""
Orchestrator Azure Function App
Handles workflow orchestration, multi-step reasoning, and agent coordination
Uses Durable Functions for long-running workflows
"""
import azure.functions as func
import azure.durable_functions as df
import logging
import json
import os
from datetime import datetime
from typing import Any, Dict, List

app = func.FunctionApp(http_auth_level=func.AuthLevel.FUNCTION)

# Configuration from environment
OPENAI_ENDPOINT = os.environ.get("AZURE_OPENAI_ENDPOINT", "")
OPENAI_KEY = os.environ.get("AZURE_OPENAI_KEY", "")
SEARCH_ENDPOINT = os.environ.get("AZURE_SEARCH_ENDPOINT", "")
SEARCH_KEY = os.environ.get("AZURE_SEARCH_KEY", "")
COSMOS_CONNECTION = os.environ.get("COSMOS_CONNECTION_STRING", "")


# Durable Functions Client
@app.route(route="orchestrate", methods=["POST"])
@app.durable_client_input(client_name="client")
async def start_orchestration(req: func.HttpRequest, client: df.DurableOrchestrationClient) -> func.HttpResponse:
    """
    Start a new orchestration workflow
    """
    logging.info("Starting new orchestration")

    try:
        req_body = req.get_json()
    except ValueError:
        return func.HttpResponse(
            json.dumps({"error": "Invalid JSON body"}),
            mimetype="application/json",
            status_code=400
        )

    workflow_type = req_body.get("workflow_type", "chat")
    payload = req_body.get("payload", {})

    # Start the appropriate orchestration
    if workflow_type == "chat":
        instance_id = await client.start_new("chat_orchestrator", None, payload)
    elif workflow_type == "document_processing":
        instance_id = await client.start_new("document_processing_orchestrator", None, payload)
    elif workflow_type == "batch_ingestion":
        instance_id = await client.start_new("batch_ingestion_orchestrator", None, payload)
    else:
        return func.HttpResponse(
            json.dumps({"error": f"Unknown workflow type: {workflow_type}"}),
            mimetype="application/json",
            status_code=400
        )

    logging.info(f"Started orchestration {instance_id}")

    return client.create_check_status_response(req, instance_id)


# Chat Orchestrator
@app.orchestration_trigger(context_name="context")
def chat_orchestrator(context: df.DurableOrchestrationContext):
    """
    Orchestrates the chat workflow:
    1. Query preprocessing
    2. Intent detection
    3. Retrieval (RAG)
    4. Response generation
    5. Post-processing
    """
    payload = context.get_input()

    # Step 1: Preprocess query
    preprocessed = yield context.call_activity("preprocess_query", payload)

    # Step 2: Detect intent
    intent = yield context.call_activity("detect_intent", preprocessed)

    # Step 3: Retrieve relevant documents
    retrieval_result = yield context.call_activity("retrieve_documents", {
        "query": preprocessed["processed_query"],
        "intent": intent,
        "filters": payload.get("filters", {})
    })

    # Step 4: Generate response
    response = yield context.call_activity("generate_response", {
        "query": preprocessed["processed_query"],
        "context": retrieval_result["documents"],
        "intent": intent,
        "conversation_history": payload.get("conversation_history", [])
    })

    # Step 5: Post-process and validate
    final_response = yield context.call_activity("postprocess_response", {
        "response": response,
        "sources": retrieval_result["documents"]
    })

    # Step 6: Save to conversation history
    yield context.call_activity("save_conversation", {
        "conversation_id": payload.get("conversation_id"),
        "user_id": payload.get("user_id"),
        "query": payload["query"],
        "response": final_response
    })

    return final_response


# Document Processing Orchestrator
@app.orchestration_trigger(context_name="context")
def document_processing_orchestrator(context: df.DurableOrchestrationContext):
    """
    Orchestrates document processing:
    1. OCR extraction
    2. Chunking
    3. Embedding generation
    4. Index update
    """
    payload = context.get_input()

    # Step 1: Extract text using Document Intelligence
    extracted = yield context.call_activity("extract_document_text", payload)

    # Step 2: Chunk the document
    chunks = yield context.call_activity("chunk_document", extracted)

    # Step 3: Generate embeddings for all chunks (fan-out)
    embedding_tasks = []
    for chunk in chunks:
        task = context.call_activity("generate_embedding", chunk)
        embedding_tasks.append(task)

    embeddings = yield context.task_all(embedding_tasks)

    # Step 4: Update search index
    result = yield context.call_activity("update_search_index", {
        "document_id": payload["document_id"],
        "chunks": chunks,
        "embeddings": embeddings
    })

    # Step 5: Update document metadata in Cosmos DB
    yield context.call_activity("update_document_metadata", {
        "document_id": payload["document_id"],
        "status": "processed",
        "chunks_count": len(chunks)
    })

    return result


# Activity Functions
@app.activity_trigger(input_name="payload")
def preprocess_query(payload: Dict[str, Any]) -> Dict[str, Any]:
    """Preprocess the user query"""
    query = payload.get("query", "")

    # Basic preprocessing
    processed = query.strip()

    return {
        "original_query": query,
        "processed_query": processed,
        "language": "en",  # TODO: Detect language
        "timestamp": datetime.utcnow().isoformat()
    }


@app.activity_trigger(input_name="payload")
def detect_intent(payload: Dict[str, Any]) -> Dict[str, Any]:
    """Detect user intent from the query"""
    query = payload.get("processed_query", "")

    # TODO: Use GPT for intent detection
    intent = {
        "primary_intent": "information_retrieval",
        "confidence": 0.85,
        "entities": [],
        "requires_rag": True
    }

    return intent


@app.activity_trigger(input_name="payload")
def retrieve_documents(payload: Dict[str, Any]) -> Dict[str, Any]:
    """Retrieve relevant documents from Azure AI Search"""
    query = payload.get("query", "")
    filters = payload.get("filters", {})

    # TODO: Implement Azure AI Search retrieval
    return {
        "documents": [],
        "total_count": 0,
        "search_score": 0.0
    }


@app.activity_trigger(input_name="payload")
def generate_response(payload: Dict[str, Any]) -> Dict[str, Any]:
    """Generate response using Azure OpenAI"""
    query = payload.get("query", "")
    context = payload.get("context", [])
    conversation_history = payload.get("conversation_history", [])

    # TODO: Implement Azure OpenAI call
    response = {
        "content": "This is a placeholder response.",
        "model": "gpt-4o",
        "tokens_used": 0,
        "grounded": True
    }

    return response


@app.activity_trigger(input_name="payload")
def postprocess_response(payload: Dict[str, Any]) -> Dict[str, Any]:
    """Post-process and validate the response"""
    response = payload.get("response", {})
    sources = payload.get("sources", [])

    # TODO: Implement validation, citation adding, etc.
    return {
        "content": response.get("content", ""),
        "sources": sources,
        "validated": True,
        "timestamp": datetime.utcnow().isoformat()
    }


@app.activity_trigger(input_name="payload")
def save_conversation(payload: Dict[str, Any]) -> Dict[str, Any]:
    """Save conversation to Cosmos DB"""
    # TODO: Implement Cosmos DB save
    return {"saved": True}


@app.activity_trigger(input_name="payload")
def extract_document_text(payload: Dict[str, Any]) -> Dict[str, Any]:
    """Extract text from document using Document Intelligence"""
    # TODO: Implement Document Intelligence OCR
    return {
        "document_id": payload.get("document_id"),
        "text": "",
        "pages": 0
    }


@app.activity_trigger(input_name="payload")
def chunk_document(payload: Dict[str, Any]) -> List[Dict[str, Any]]:
    """Chunk document into smaller pieces"""
    # TODO: Implement semantic chunking
    return []


@app.activity_trigger(input_name="payload")
def generate_embedding(payload: Dict[str, Any]) -> Dict[str, Any]:
    """Generate embedding for a chunk"""
    # TODO: Implement embedding generation
    return {
        "chunk_id": payload.get("chunk_id"),
        "embedding": []
    }


@app.activity_trigger(input_name="payload")
def update_search_index(payload: Dict[str, Any]) -> Dict[str, Any]:
    """Update Azure AI Search index"""
    # TODO: Implement index update
    return {"indexed": True}


@app.activity_trigger(input_name="payload")
def update_document_metadata(payload: Dict[str, Any]) -> Dict[str, Any]:
    """Update document metadata in Cosmos DB"""
    # TODO: Implement metadata update
    return {"updated": True}
